Scrollable UIs in Dart with SingleChildScrollView
In Dart, SingleChildScrollView is a widget that allows you to create a scrollable area within your user interface. It's useful when you have content that doesn't fit within the bounds of the device screen and you want users to be able to scroll through it.

The SingleChildScrollView widget contains a single child widget that can be larger than the size of the screen. When the content of the child widget is too large to fit on the screen, the widget automatically enables scrolling.

The widget can be customized to suit your specific design requirements. You can define properties such as the scroll direction (vertical or horizontal), the padding around the content, and the physics used to scroll the content.