Customizable Lists with ListView.Separated Flutter
ListView.Separated is a widget in the Flutter framework that allows you to create a scrollable list of items with custom separators between them. It's similar to ListView.Builder, but instead of creating a single widget for each item in the list, it allows you to define two widgets - one for the item and one for the separator. This gives you more control over the appearance of the list, and allows you to create more complex layouts.

With ListView.Separated, you can easily create a list of items with custom separators, such as horizontal dividers, vertical dividers, or any other widget that you want to use as a separator. You can also customize the appearance of each item in the list using the itemBuilder parameter, and customize the appearance of each separator using the separatorBuilder parameter.