Flutter Padding Widget: Adding Space
The Padding widget in Dart's Flutter framework allows developers to add space around child widgets. This widget is useful for creating visual separation between widgets, adjusting the layout, and improving the overall look and feel of the user interface. By specifying the amount of padding to add, developers can create a more polished and professional-looking design.