Margin Widget: Adjusting Widget Positioning
This topic focuses on the Margin widget in Dart's Flutter framework, which allows developers to adjust the positioning of child widgets within a container. The Margin widget adds space around the outside of a container, creating a visual separation between the container and other widgets. By adjusting the amount of margin, developers can control the layout and appearance of their user interface. The Margin widget is a useful tool for creating well-designed and visually appealing user interfaces.